Clarify branch and patch naming for the openela-contrib
The rationale for the change is that openela-main only contains
machine-generated branches that do not accept contributions. All
contributions require a instantiation of the same package or of
a new package (with a maintainer) in openela-contrib, and to ease
rebasing on top of openela-main, the branch names need to carry
a -contrib suffix.